+++ tango/integrations/torch/train.py
@@ -390,6 +390,8 @@ def _train(
         for step, batch in train_batch_iterator:
 
             def save_state():
+                train_batch_iterator.set_description(desc="Training (saving checkpoint)")
+
                 state_path_for_step = work_dir / f"state_worker{worker_id}_step{step + 1}.pt"
                 temp_state_file = tempfile.NamedTemporaryFile(
                     "w+b", dir=work_dir, delete=False, suffix=".pt"
@@ -427,6 +429,9 @@ def _train(
                         if best_state_path.is_symlink():
                             best_state_path.unlink()
                         best_state_path.symlink_to(state_path_for_step)
+
+                    # Reset progress desc.
+                    train_batch_iterator.set_description(desc="Training")
                 finally:
                     if os.path.exists(temp_state_file.name):
                         os.remove(temp_state_file.name)
@@ -488,7 +493,7 @@ def _train(
 
             # Gather average loss across all workers.
             if (should_log_this_step or should_validate_this_step) and is_distributed:
-                batch_loss_tensor = torch.tensor(batch_loss)
+                batch_loss_tensor = torch.tensor(batch_loss, device=device)
                 dist.all_reduce(batch_loss_tensor)
                 batch_loss = batch_loss_tensor.item() / world_size
 
@@ -538,7 +543,7 @@ def _train(
 
                         # Average loss across all workers.
                         if is_distributed and should_log_this_step:
-                            val_loss_tensor = torch.tensor(val_loss)
+                            val_loss_tensor = torch.tensor(val_loss, device=device)
                             dist.all_reduce(val_loss_tensor)
                             val_loss = val_loss_tensor.item() / world_size
 
@@ -570,6 +575,9 @@ def _train(
             # Checkpoint.
             if should_checkpoint_this_step:
                 save_state()
+
+    if is_distributed:
+        dist.barrier()
 
     if not is_distributed:
         return model