+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed `DirectoryStepCache` to `LocalStepCache` (registered as "local").
 
+### Fixed
+
+- Fixed bug with `TorchTrainStep` during distributed training.
+
 ## [v0.1.0](https://github.com/allenai/tango/releases/tag/v0.1.0) - 2021-10-11
 
 ### Added