+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.1.1](https://github.com/allenai/tango/releases/tag/v0.1.1) - 2021-10-12
+
 ### Added
 
 - `TorchTrainStep` now displays a progress bar while saving a checkpoint to file.